pyproject.toml
[tool.poetry]
name = "langchain-llm-app"
version = "0.1.0"
description = ""
authors = ["samoylovartem <samoylovartem07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.352"
openai = "^1.6.1"
streamlit = "^1.29.0"
python-dotenv = "^1.0.0"
wikipedia = "^1.4.0"
numexpr = "^2.8.8"
youtube-transcript-api = "^0.6.1"
tiktoken = "^0.5.2"
faiss-cpu = "^1.7.4"
langdetect = "^1.0.9"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pre-commit = "^3.6.0"
isort = "^5.13.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
